Make traits unnecessarily robust
Using the checked methods in the traits is probably not needed. I couldn't create ub with the unchecked methods or a panic with the checked methods. That said, there may be way if someone else uses unsafe code. I made them checked here just in case, since I don't see a real performance reason to make them not. Since it just has an extra boolean check.